Dev: move command palette out of covid
