+++ charts/CommandPalette.tsx
