+++ charts/covidDataExplorer/CovidConstants.ts
@@ -104,3 +104,22 @@ export const metricLabels: { [key in MetricKind]: string } = {
     tests_per_case: "Tests per confirmed case",
     positive_test_rate: "Share of positive tests"
 }
+
+// todo: auto import from covid repo.
+export const covidAnnotations = `location,date,cases_annotations,deaths_annotations
+Spain,2020-04-19,methodology change,
+Spain,2020-04-25,methodology change,methodology change
+Ecuador,2020-05-08,methodology change,
+United Kingdom,2020-05-20,methodology change,
+France,2020-06-02,methodology change,
+India,2020-06-17,,earlier deaths added
+Chile,2020-06-18,earlier cases added,
+Italy,2020-06-25,,methodology change
+United States,2020-06-26,,probable/earlier deaths added
+United States,2020-07-01,,probable/earlier deaths added
+United Kingdom,2020-07-03,methodology change,
+Czech Republic,2020-07-05,,methodology change
+Kyrgyzstan,2020-07-18,methodology change,methodology change
+Chile,2020-07-18,,methodology change
+Peru,2020-07-24,,earlier deaths added
+European Union,,Some EU countries changed methodology. See country-by-country series.,Some EU countries changed methodology. See country-by-country series.`
